feat: home component cleaned up and move form to the modal box
