+++ src/app/home/home.component.html
@@ -18,7 +18,7 @@
   </div>
 
   <div class="modal-body">
-    <div class="container">Hello World</div>
+    <app-add-beneficiary></app-add-beneficiary>
   </div>
   <div class="modal-footer">
     <button
@@ -53,68 +53,9 @@
 </div>
 
 <div *ngIf="!showReviewPage">
-  <div class="card">
-    <div class="card-body">
-      <h4 class="modal-title" id="modal-basic-title mb-1">
-        &lt;Account&gt; is missing a primary beneficiary
-      </h4>
-      <p class="fnt-09 mt-1">
-        Now that you&#039;ve transferred money to this account, let&#039;s make
-        sure you secure your assets with a beneficiary. Naming a beneficiary is
-        free and it only takes a few minutes.
-      </p>
-
-      <form [formGroup]="beneficiariesForm" (ngSubmit)="onSubmit()" novalidate>
-        <div class="beneficiary-form" formArrayName="beneficiaries">
-          <div *ngFor="let beneficiary of getArrayElements(); let i = index">
-            <div
-              [formGroupName]="i"
-              class="pt-3"
-              [ngClass]="{ 'border-top': i > 0 }"
-            >
-              <div
-                class="d-flex justify-content-end position-absolute right-1 gap-2"
-              >
-                <button
-                  type="button"
-                  class="btn btn-light"
-                  *ngIf="
-                    beneficiaries.length < 2 && i == beneficiaries.length - 1
-                  "
-                  (click)="addNewBeneficiary()"
-                >
-                  <i class="bi bi-plus-circle"></i> Add New
-                </button>
-                <button
-                  type="button"
-                  class="btn btn-warning"
-                  *ngIf="beneficiaries.length > 1"
-                  (click)="removeNewBeneficiary(i)"
-                >
-                  <i class="bi bi-trash"></i> Remove
-                </button>
-              </div>
-
-              <app-beneficiary
-                [items]="beneficiary"
-                [fieldIndex]="i"
-              ></app-beneficiary>
-            </div>
-          </div>
-        </div>
-
-        <div class="row border-top pt-2">
-          <div class="col text-center">
-            <button type="submit" class="btn btn-primary">
-              <i class="bi bi-box-arrow-in-right"></i> Next
-            </button>
-          </div>
-        </div>
-      </form>
-    </div>
-  </div>
+  
 </div>
 
 <div *ngIf="showReviewPage">
-  <app-review-page [reviewData]="getReviewData()" (backButtonClicked)="toggleReviewPage(false)"></app-review-page>
+  <!-- <app-review-page [reviewData]="getReviewData()" (backButtonClicked)="toggleReviewPage(false)"></app-review-page> -->
 </div>
