+++ src/styles.scss
@@ -11,3 +11,7 @@
 .pl-12 {
     padding-left: 12rem;
 }
+
+.right-1 {
+    right: 1rem;
+}
